end of day
Added extension class for HttpClient, coping over code from DotNetCoreExam GitHub project. Added a method to allow getting via a query string in the url.

Updated GenericFactory so that user can chose to go New to Old or Old to New. Old to New is needed to get a new reference in some cases, and is the default way the method runs. The New to Old way is needed when an existing reference simply needs updating.