end of day
Created/Copied Persistance specific code from DotNetExam project on github, and improved it for future use. now it makes use of generic types. Also renamed the interface taken from there to IHandler.

Created/Copied Startup specific code from DotNetExam project on github, and improved it for future use. Is cabable of taking in any number of modules, as long as they implement IStartupModule. Currently have created two modules. One is EntityFrameworkStartupModule, which take the type of dbcontext and the type of the implementation of IHandler, setting up dependency injection for these. Have also made a Swagger module called SwaggerStartupModule, which allows easy enabling/adding of Swagger to Api's.

oh and also removed XmlManipulator, as it was not not fully developed anyways, and so should never be used anyways. Making use of Json is also a better way to transport data, which has more support.