end of day
Made Handler methods return the same version, but with a Task, indicating that it should be an asynchronous action.

Introduced an AdvancedController class and interface, that has some default code for Get, Post, Put and Delete, using IEntity as type constraints, this should mean that it is only a matter of inheriting at the correct place for a controllers methods to already exist.