Uploading as a mayor breakthrough has been archived.
Added "Await" type to LogType. used for areas that are waiting on a condition, for example TaskExtensions, in WaitUntil and WaitWhile.

Updated WaitUntil and WaitWhile to allow log messages whenever it checks its condition. A very slow frequency of checking is recommended, as otherwise the log becomes almost unreadable.

Added NegotiateInfo, which matches the Json response from a Azure Function where a user can Negotiate for connection information to a Azure SignalR.

Updated HubProxy, with a bunch of features.
 - Added Boolean input to indicate if there should be logged when waiting on a specific connection state.
 - Added a public get for the HubConnectionState, so user of the hub, can make decisions depending on the state.
 - Added event for when the state becomes Connecting or Reconnecting.
 - Split Initialization into two directions, one for Self-Hosted hubs, and one for Azure Hosted ones. Choseing which type is done via a simple Boolean input to Init, which default to false, meaning Self-Hosted is the default.