Worked on improving BaseHandler
Marked Add, Update, and AddMultiple Obsolete
Created AddAndRetrieve, which is the new Add, that also retrieves the newly added entity
Created UpdateAndRetrieve, which is the new Update, that also retrieves the newly updated entity.
Created AddMultipleAndRetrieve, which is the new AddMultiple, that also retrieves the newly added entities
Created TypeExtensions for a helper method that retrieves MethodInfo based on specified Type and method name.
Updated BaseContext to overwrite all version of SaveChanges, both async and sync versions.
Removed Save from IHandler, as it is only a helper method on BaseHandler, that is always called, between adding/updating and retrieving.