Continued work on TimeTracker.
Created BasePropertyChanged, which contains base code for implementing INotifyPropertyChanged.
Added comment to BaseEntity, noting that it should be reworked into a dependecy injection based module.
Added LICENSE.md containing the license for Extended.Wpf.Toolkit, which will be used to create some Wpf components when i get to creating the client.
Updated text written by SetCaller() to specify what it was before.
Created ChangedEntity, which is a further expansion on BasePropertyChanged for IEntities.
Changed NLogStartupModule from making a singleton LoggerManager to a Scoped one, as different controllers , should have their own, so the calls to LoggerManager can be tracked better.
Updated AdvancedController and IAdvancedController to call all parameters entity instead of data.