Created ConditionalDateTimePicker and ImpossibleException. Also Worked on NavigationWindow
ImpossibleException is meant to be thrown from areas that should be unreachable in theory. An example of this is the default of a switch using a boolean to switch on.
ConditionalDateTimePicker allows up to two dates to be picked, with a dropdown to enable/disable some or all of the DateTimePicker.
NavigationWindow is a implementation of a hamburger navigation for WPF, done from scratch. Could find any when googleing, so i chose to work on building my own.

Neither has been tests, if they are visually correct - yet.